Botany Splicer Infusion Process Rework (#6910)
## About The Pull Request

This pull request changes the infusion process of the Seed Splicer
machine in Botany. The PR introduces a chance-based system where players
can add a beaker filled with any reagent (With the Synthesizable Tag)
inside of the Splicer and have a chance to infuse a gene of that reagent
in the seed (Where every 5 units is a 1% chance to infuse that gene into
the seed). If the process is successful the gene will be able to produce
that reagent at a rate between 3%-25%. Trying to infuse a seed with a
gene that already produces that reagent will just increase the rate it
produces that chemical.

The Infusion process will also cause infusion damage to the seed (Where
every 1 unit is 0.1 Infusion Damage). Once a seed reaches 100 damage it
will disintegrate.

Also regarding Infusion mutations, in order to obtain a plant that
requires an infusion mutation, you'll need to successfully infuse the
required chemical into the plant in order to unlock that mutation. (EX:
Gunpowder for Coconut Guns, Eggnog for Egg-Plants, etc.)

## Why It's Good For The Game

This change makes it so botany players are more of an important role
around the station. Currently, the way a lot of botany players will play
is that they'll try to make a bunch of lesser strange seeds and they'll
choose which reagent genes in those seeds they want to extract and
depending on those genes help them decide which department they want to
help (most of the time they decide to choose medical). With this change,
people are more likely to ask botany's help if they want to produce a
tedious chemical to make. Some examples I can include are if the Chef
wants Botany to mass-produce a certain type of soup they have, the
Bartenders need to mass-produce a hard-to-make drink, or if medical
needs a supply of Oxandrolone using the beakers they already have.

The infusion damage change also allows us to experiment more with
infusion mutations as since they're now harder to obtain, this'll allow
us to make them more powerful in the future or lock some already
powerful mutations behind them. For example, locking Holymelons behind
Holy water would make it so the Chaplain's help is required if Botany
players need to make them. Locking Bluespace fruits behind Bluespace
Dust would also make it extremely difficult, if not impossible, for
prisoners to instantly make them in order to escape.

## Changelog

:cl:
add: The Seed Splicer's infusion process has been overhauled.
Synthesizable Reagents from a beaker can now be infused into a seed one
by one. Each reagent in the beaker has a chance to be infused into the
seed, with the chance increasing based on the reagent's volume (1% per 5
units).
balance: Infusing seeds now causes damage. The total volume of reagents
in the beaker will inflict a base amount of damage to the seed (0.1
damage per unit of reagent). Seeds can be destroyed if they take too
much damage.
/:cl:

---------

+++ monkestation/code/modules/hydroponics/machines/splicer.dm
@@ -261,13 +261,20 @@
 	potential_damage = stats["damage"]
 
 /obj/machinery/splicer/proc/infuse()
-	if(!held_beaker)
+	if(!held_beaker || !seed_1) /// Checks if we have a beaker and a seed to infuse
 		return
-	seed_1.infusion_damage += potential_damage
-	if(seed_1.infusion_damage >= 100)
+	potential_damage = held_beaker.reagents.total_volume / 10 /// Every 10 units of reagents in the beaker will cause 1 damage to the seed.
+	seed_1.infusion_damage += min(potential_damage, 100 - seed_1.infusion_damage)
+
+	if(seed_1.infusion_damage >= 100) /// If the seed has taken too much damage, it gets deleted.
+		to_chat(usr, span_warning("[seed_1] has become too damaged from infusion and disintegrated!"))
 		qdel(seed_1)
 		seed_1 = null
+		stats = list()
+		potential_damage = 0
 		return
+
+	var/list/successful_reagents = list()
 
 	seed_1.adjust_potency(stats["potency_change"])
 	seed_1.adjust_yield(stats["yield_change"])
@@ -278,7 +285,65 @@
 	seed_1.adjust_weed_rate(stats["weed_rate_change"])
 	seed_1.adjust_maturation(stats["maturation_change"])
 
-	seed_1.check_infusions(held_beaker.reagents.reagent_list)
-	held_beaker.reagents.remove_all(held_beaker.reagents.total_volume)
+	if(held_beaker.reagents && held_beaker.reagents.reagent_list.len > 0)
+		var/list/current_reagent_instances = held_beaker.reagents.reagent_list.Copy()
+
+		for (var/datum/reagent/reagent_instance in current_reagent_instances)
+			if(!reagent_instance)
+				continue
+
+			var/reagent_id_path = reagent_instance.type
+			if (!(reagent_instance.chemical_flags & REAGENT_CAN_BE_SYNTHESIZED))
+				to_chat(usr, span_warning("[reagent_instance.name] cannot be infused into plants!"))
+				continue
+
+			var/volume = reagent_instance.volume
+			if (volume <= 0)
+				continue
+
+			var/infusion_chance = min(floor(volume / 5), 100)
+
+			if(prob(infusion_chance))
+				var/random_rate = rand(3, 25) / 100
+
+				var/datum/plant_gene/reagent/existing_gene = null
+				for(var/datum/plant_gene/gene_check in seed_1.genes)
+					if(istype(gene_check, /datum/plant_gene/reagent))
+						var/datum/plant_gene/reagent/reagent_gene = gene_check
+						if(reagent_gene.reagent_id == reagent_id_path)
+							existing_gene = reagent_gene
+							break
+
+				if(existing_gene)
+					/// Add to existing gene's rate
+					to_chat(usr, span_notice("Increased [reagent_instance.name] rate in [seed_1] from [round(existing_gene.rate * 100)]% to [round(existing_gene.rate * 100) + round(random_rate * 100)]%."))
+					existing_gene.rate += random_rate
+				else
+					/// Create a new gene with the random rate
+					var/datum/plant_gene/reagent/new_gene = new /datum/plant_gene/reagent(reagent_id_path, random_rate)
+					if(new_gene.can_add(seed_1))
+						seed_1.genes += new_gene
+						to_chat(usr, span_notice("Successfully infused [reagent_instance.name] into [seed_1] with a rate of [round(new_gene.rate * 100)]%."))
+					else
+						/// Skips adding the gene mutation if it failed to add
+						to_chat(usr, span_warning("Could not add new gene for [reagent_instance.name] to [seed_1]."))
+						qdel(new_gene)
+						continue
+
+				successful_reagents += reagent_instance
+			else
+				to_chat(usr, span_notice("Attempted to infuse [reagent_instance.name] into [seed_1], but it failed."))
+
+
+		seed_1.reagents_from_genes()
+	else
+		to_chat(usr, span_warning("The beaker is empty! Nothing to infuse."))
+
+	seed_1.check_infusions(successful_reagents)
+	if(held_beaker && held_beaker.reagents)
+		held_beaker.reagents.remove_all(held_beaker.reagents.total_volume)
+	successful_reagents = list()
 	stats = list()
 	potential_damage = 0
+
+	to_chat(usr, span_notice("[seed_1] infusion process complete."))
